Invalid input additions
The only criticism I have to offer is that at times it was difficult to know whether I was entering a valid input. I fixed this specifically for when you build a team --- if the user wants to add a player and they enter a position that doesn't exist, it will simply ask you to try again. I did the same for the remove player as well. Makes it easier for people who don't know too much about soccer. Other than that, couldn't find any bugs. Cool project!